+++ config.yaml.defaults
@@ -625,6 +625,7 @@ user:
 
 ports:
   facility_queue: 64510
+  notification_queue: 64610
 
 gcn:
   server: gcn.nasa.gov
