+++ services/notification_queue/supervisor.conf
@@ -1,0 +1,5 @@
+[program:notification_queue]
+command=/usr/bin/env python services/notification_queue/notification_queue.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+stdout_logfile=log/notification.log
+redirect_stderr=true
